Add launch configurations
This package can now be launched. Something to note is that the zed_wrapper
uses nodelets that put all of the pub/sub topics on the private NodeHandle.
This means all topics are within the zed_node namespace (or whatever name the
node is given). This makes remapping a bit more annoying because the
c1t_zed_driver needs to also subscribe to the same topic (meaning it needs to
know the zed node's name). Otherwise, remapping will be more tedious.
Therefore, the c1t_zed_driver was given a parameter for the zed node name.

c1t_zed_driver.cpp
* add parameter for zed node name so node can subscribe to proper topic

c1t_zed_driver.launch
* add left camera topic remapping so it can be used with CARMA
* add c1t_zed_driver node call
* add custom zed_node launch file call

zed_node.launch
* add zed_camera launch file include without a namespace. The original launch
  file from the zed_wrapper package wrapped the zed node in a namespace, which
  made remapping even more annoying. This helps remedy some of those issues.